Implemented help as its own cog -This means that help now has its own subcommands -Help commands were deprecated from other cogs in exchange
