+++ cogs/OH_queue.py
@@ -117,40 +117,6 @@ class OH_Queue(commands.Cog):
     async def queue(self, context: Context):
         pass
 
-    # This is a subcommand of queue
-    # invoked by >queue help
-    # you may want to move this back to a cog command instead of a subcommand
-    @queue.command()
-    async def help(self, context: Context):
-        embed = discord.Embed(title="Help", description="", color=0x7289da)
-        embed.set_author(
-            name="OH Bot",
-            url="https://discordbots.org/bot/472911936951156740",
-            icon_url="https://i.imgur.com/i7vvOo5.png"
-        )
-
-        embed.add_field(
-            name=f'**Commands**',
-            value=f'**Enter the OH queue by using the following command:**\n\n`>enterqueue or >eq`\n\n------------\n\n'
-                  f'**Leave the OH queue by using the following command:**\n\n`>leavequeue or >lq`\n\n------------\n\n',
-            inline=False
-        )
-        if isAdmin(context):
-            adminEmbed = discord.Embed(title="Admin Help", description="", color=0x7289da)
-            adminEmbed.set_author(
-                name="OH Bot",
-                url="https://discordbots.org/bot/472911936951156740",
-                icon_url="https://i.imgur.com/i7vvOo5.png"
-            )
-            adminEmbed.add_field(
-                name=f"**Admin only commands",
-                value=f'**Clear the OH queue by using the following command:**\n\n`>leavequeue or >lq`\n\n------------\n\n'
-                      f'**Call a student into your OH session by using the following command:**\n\n`>dequeue or >dq`\n\n------------\n\n',
-                inline=False
-            )
-        await getSender(context).send(embed=embed)
-        await selfClean(context)
-
 
 def setup(client):
     """
